new defender test

+++ lib/Lowlevel/Line.h
@@ -8,10 +8,10 @@ class LineC {
         int VW_Summe;
         int Summe;
         float calib_line[40];
-        uint16_t line[32] ;
-        uint16_t lineVW[8];
-        uint16_t Values_raw[32];
-        uint16_t Values_raw_VW[8];
+        uint8_t line[32] ;
+        uint8_t lineVW[8];
+        int Values_raw[32];
+        int Values_raw_VW[8];
         void read_Fast(); 
         void init();
         void Calibrate(int Modes);
